Implenting model, test window, and trial data saving
Implemented segmenting function for offline and online data to create windows to be passed to the model for retraining. Windows are defaulted to be 125 samples, or 1 second of data on the real board.
Implemented saving of the random test windows, as well as saving of the raw batch data (for whatever purposes) and the split up batch data that's put in the queue.